GeneralTensorFlowNetwork extends TensorFlowArchitecture. We should consider changing this in the future
